requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
jinja2>=3.1.0
pyyaml>=6.0
